catseg/predict/catapi/requirements.txt
fastapi[standard]
mangum
numpy
pillow
ultralytics
pydantic